wlan: Unregister only matching devices
Userspace is currently maintaining the wwan and wlan device per link and
per ip family and generates reference counts based on these states. When
a new network is created, operations based on the reference counts may
override the metadata in kernel as kernel only supports one type of
device in total.

A sample sequence of operations can be the following -

1. LBO teardown on old WLAN link
2. LBO setup on new WLAN link
3. WLAN teardown on old WLAN link
4. WLAN setup on new WLAN link

When step 3 occurs, the reference count for old WLAN link would require
the cleanup of WLAN device, however the WLAN device was already setup on
the new WLAN link. As a result, this would cause the WLAN device
information to be removed and step 4 can't add it back as the reference
count is already too high.

This change now adds matching of the device prior to removal. Adding
full support for multiple WLAN devices would need a more invasive
change.

CRs-Fixed: 4161739
Change-Id: I76855bf2344b8830d94e614df3224667235f48e1
Signed-off-by: Subash Abhinov Kasiviswanathan <subash.a.kasiviswanathan@oss.qualcomm.com>